将存储引擎改回storm，达到 免安装，简化使用

+++ core/tunnel.go
@@ -11,7 +11,7 @@ type Tunnel interface {
 	Open() error
 	Close() error
 	GetTunnel() *models.Tunnel
-	GetLink(id int64) (*Link, error)
+	GetLink(id int) (*Link, error)
 }
 
 type baseTunnel struct {
